Refactor: Consolidate agent registration logic
Move agent registration and validation logic from agent-interop.ts to agnet.ts to centralize agent management.
